Add caching

+++ install-pcre.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 
-mkdir pcre
+mkdir -p pcre
 cd pcre
-wget "https://ftp.pcre.org/pub/pcre/pcre-${PCRE}.tar.gz"
-tar xvfz "pcre-${PCRE}.tar.gz"
+[ ! -e "pcre-${PCRE}.tar.gz" ] && wget "https://ftp.pcre.org/pub/pcre/pcre-${PCRE}.tar.gz"
+[ ! -e "pcre-${PCRE}" ] && tar xvfz "pcre-${PCRE}.tar.gz"
 cd "pcre-${PCRE}"
-./configure
-make
-cp .libs/* .
+[ ! -e Makefile ] && ./configure
+[ ! -e pcretest ] && make
+[ ! -e libpcre.a ] && cp .libs/* .
 cd ../..
